add Modal component and welcome text, change spinner-style
